# secret-share-client/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SecretShare API",
    "version": "1.0.0",
    "description": "API for SecretShare"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "API server"
    }
  ],
  "paths": {
    "/api/secrets/": {
      "post": {
        "summary": "Create a new secret",
        "tags": [
          "Secrets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateSecretRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Secret created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ICreateSecretResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/secrets/defaults": {
      "get": {
        "summary": "Get the default settings for secrets",
        "tags": [
          "Secrets"
        ],
        "responses": {
          "200": {
            "description": "Default secret settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SecretDefaults"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/secrets/{identifier}": {
      "get": {
        "summary": "Retrieve a secret by its identifier",
        "tags": [
          "Secrets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The identifier of the secret"
          },
          {
            "in": "header",
            "name": "x-secret-password",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The secondary secret password (passed as a request header to avoid logging in server access logs)"
          }
        ],
        "responses": {
          "200": {
            "description": "Secret retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "secret": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (expired, view limit reached, IP not allowed, or invalid secondary secret)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ForbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Secret not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a given secret",
        "description": "Delete a secret by its creator identifier. This action is irreversible.",
        "tags": [
          "Secrets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The creator identifier of the secret"
          }
        ],
        "responses": {
          "200": {
            "description": "Secret deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Secret not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/secrets/logs/{identifier}": {
      "get": {
        "summary": "Get the access logs for a secret",
        "description": "Retrieve the access logs for a secret by its creator identifier",
        "tags": [
          "Secrets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The creator identifier of the secret"
          }
        ],
        "responses": {
          "200": {
            "description": "Access logs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "logs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ISecretAccessLog"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Secret not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/secrets/stats/{creatorIdentifier}": {
      "get": {
        "summary": "Get usage statistics for a secret",
        "description": "Retrieve dashboard-ready usage statistics for a secret by its creator identifier",
        "tags": [
          "Secrets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "creatorIdentifier",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The creator identifier of the secret"
          }
        ],
        "responses": {
          "200": {
            "description": "Secret statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/GetSecretStatsResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Secret not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Check the status of all services",
        "description": "Returns the status of all services and indicates if all are running.",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "All services are running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "services": {
                      "type": "object",
                      "properties": {
                        "MongoConnection": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "RUNNING"
                        },
                        "RabbitMQConnection": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "RUNNING"
                        },
                        "ExpressServer": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "RUNNING"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Some or all services are not running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "services": {
                      "type": "object",
                      "properties": {
                        "MongoConnection": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "UNKNOWN"
                        },
                        "RabbitMQConnection": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "UNKNOWN"
                        },
                        "ExpressServer": {
                          "type": "string",
                          "enum": [
                            "STARTING",
                            "RUNNING",
                            "STOPPING",
                            "STOPPED",
                            "ERROR",
                            "UNKNOWN"
                          ],
                          "example": "UNKNOWN"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ICreateSecretRequest": {
        "type": "object",
        "properties": {
          "encryptedSecret": {
            "type": "string"
          },
          "ipRestrictions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxViews": {
            "type": "number"
          },
          "secretPassword": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "emailNotification": {
            "type": "string"
          }
        }
      },
      "ICreateSecretResponse": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string"
          },
          "creatorIdentifier": {
            "type": "string"
          }
        }
      },
      "SecretDefaults": {
        "type": "object",
        "properties": {
          "maxViews": {
            "type": "number"
          },
          "defaultExpirationLength": {
            "type": "number"
          }
        }
      },
      "ISecretAccessLog": {
        "type": "object",
        "properties": {
          "ipAddress": {
            "type": "string"
          },
          "accessDate": {
            "type": "string",
            "format": "date-time"
          },
          "accessGranted": {
            "type": "boolean"
          },
          "referrer": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "requestHeaders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requestBody": {
            "type": "string"
          }
        }
      },
      "GetSecretStatsResponse": {
        "type": "object",
        "properties": {
          "reportedViews": {
            "type": "number"
          },
          "totalViews": {
            "type": "number"
          },
          "uniqueViews": {
            "type": "number"
          },
          "grantedAttempts": {
            "type": "number"
          },
          "refusedAttempts": {
            "type": "number"
          },
          "maxViews": {
            "type": "number",
            "nullable": true,
            "description": "null indicates unlimited views"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hasPassword": {
            "type": "boolean"
          },
          "ipRestrictions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emailNotification": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "expired",
              "exhausted"
            ]
          }
        }
      }
    },
    "responses": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "NotFoundResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Secret not found"
          }
        }
      },
      "ForbiddenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Forbidden"
          }
        }
      },
      "InternalServerErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}