# secret-share-server/src/routes/secretRoutes.ts
import express, { Request, Response } from "express";
import rateLimit from "express-rate-limit";
import { Types } from "mongoose";
import Secret from "@models/Secret";
import { v4 as uuidv4 } from "uuid";
import { SecretDefaults } from "@models/SecretDefaults";
import { ICreateSecretRequest } from "@models/requests";
import {
  IApiResponse,
  ICreateSecretResponse,
  IGetSecretResponse,
  IDeleteSecretResponse,
  IGetSecretLogsResponse,
  IGetSecretStatsResponse,
} from "@models/responses";
import { validateSecret } from "../validators/secretValidator";

const router = express.Router();

// General rate limiter: 100 requests per 15 minutes per IP
const generalLimiter = rateLimit({
  windowMs: 15 * 60 * 1000,
  max: 100,
  standardHeaders: true,
  legacyHeaders: false,
  message: { success: false, error: "Too many requests, please try again later." },
});

// Strict rate limiter for sensitive endpoints (create/retrieve secrets): 20 per 15 minutes per IP
const strictLimiter = rateLimit({
  windowMs: 15 * 60 * 1000,
  max: 20,
  standardHeaders: true,
  legacyHeaders: false,
  message: { success: false, error: "Too many requests, please try again later." },
});

router.use(generalLimiter);

/**
 * @swagger
 * components:
 *   schemas:
 *     ICreateSecretRequest:
 *       type: object
 *       properties:
 *         encryptedSecret:
 *           type: string
 *         ipRestrictions:
 *           type: array
 *           items:
 *             type: string
 *         maxViews:
 *           type: number
 *         secretPassword:
 *           type: string
 *         expirationDate:
 *           type: string
 *           format: date-time
 *         emailNotification:
 *           type: string
 *     ICreateSecretResponse:
 *       type: object
 *       properties:
 *         identifier:
 *           type: string
 *         creatorIdentifier:
 *           type: string
 *     SecretDefaults:
 *       type: object
 *       properties:
 *         maxViews:
 *           type: number
 *         defaultExpirationLength:
 *           type: number
 *     ISecretAccessLog:
 *       type: object
 *       properties:
 *         ipAddress:
 *           type: string
 *         accessDate:
 *           type: string
 *           format: date-time
 *         accessGranted:
 *           type: boolean
 *         referrer:
 *           type: string
 *         userAgent:
 *           type: string
 *         requestHeaders:
 *           type: array
 *           items:
 *             type: string
 *         requestBody:
 *           type: string
 *     GetSecretStatsResponse:
 *       type: object
 *       properties:
 *         reportedViews:
 *           type: number
 *         totalViews:
 *           type: number
 *         uniqueViews:
 *           type: number
 *         grantedAttempts:
 *           type: number
 *         refusedAttempts:
 *           type: number
 *         maxViews:
 *           type: number
 *           nullable: true
 *           description: null indicates unlimited views
 *         expirationDate:
 *           type: string
 *           format: date-time
 *           nullable: true
 *         createdAt:
 *           type: string
 *           format: date-time
 *           nullable: true
 *         hasPassword:
 *           type: boolean
 *         ipRestrictions:
 *           type: array
 *           items:
 *             type: string
 *         emailNotification:
 *           type: string
 *           nullable: true
 *         status:
 *           type: string
 *           enum: [active, expired, exhausted]
 *   responses:
 *     SuccessResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: true
 *         data:
 *           type: object
 *     ErrorResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: false
 *         error:
 *           type: string
 *     NotFoundResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: false
 *         error:
 *           type: string
 *           example: Secret not found
 *     ForbiddenResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: false
 *         error:
 *           type: string
 *           example: Forbidden
 *     InternalServerErrorResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: false
 *         error:
 *           type: string
 *           example: Internal server error
 */

/**
 * @swagger
 * /api/secrets/:
 *   post:
 *     summary: Create a new secret
 *     tags: [Secrets]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/ICreateSecretRequest'
 *     responses:
 *       201:
 *         description: Secret created successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   $ref: '#/components/schemas/ICreateSecretResponse'
 *       400:
 *         description: Bad request
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/ErrorResponse'
 */
router.post(
  "/",
  strictLimiter,
  async (
    req: Request<{}, {}, ICreateSecretRequest>,
    res: Response<IApiResponse<ICreateSecretResponse>>
  ): Promise<void> => {
    const {
      encryptedSecret: encryptedSecret,
      ipRestrictions,
      maxViews,
      secretPassword,
      expirationDate,
      emailNotification,
    } = req.body;

    const identifier = uuidv4().replace(/-/g, "").substring(0, 25);
    const creatorIdentifier = uuidv4().replace(/-/g, "").substring(0, 25);
    const secretDefaults = new SecretDefaults();
    const ipRestrictionsSet = new Set((ipRestrictions ?? []).filter((ip) => !!ip));

    // TODO: We should consider adding an additional encryption step to the encryptedSecret
    const password = new Secret({
      identifier,
      encryptedSecret: encryptedSecret,
      creatorIdentifier,
      ipRestrictions: Array.from(ipRestrictionsSet),
      maxViews: maxViews || secretDefaults.maxViews,
      secretPassword,
      expirationDate:
        expirationDate ||
        new Date(Date.now() + secretDefaults.defaultExpirationLength),
      emailNotification,
    });

    try {
      await password.save();
      res
        .status(201)
        .send({ success: true, data: { identifier, creatorIdentifier } });
    } catch (error: any) {
      res.status(400).send({ success: false, error: error });
    }
  }
);

/**
 * @swagger
 * /api/secrets/defaults:
 *   get:
 *     summary: Get the default settings for secrets
 *     tags: [Secrets]
 *     responses:
 *       200:
 *         description: Default secret settings retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   $ref: '#/components/schemas/SecretDefaults'
 */
router.get(
  "/defaults",
  async (_: Request, res: Response<IApiResponse<SecretDefaults>>) => {
    const secretDefaults = new SecretDefaults();
    res.send({ success: true, data: secretDefaults });
  }
);

/**
 * @swagger
 * /api/secrets/{identifier}:
 *   get:
 *     summary: Retrieve a secret by its identifier
 *     tags: [Secrets]
 *     parameters:
 *       - in: path
 *         name: identifier
 *         schema:
 *           type: string
 *         required: true
 *         description: The identifier of the secret
 *       - in: header
 *         name: x-secret-password
 *         schema:
 *           type: string
 *         required: false
 *         description: The secondary secret password (passed as a request header to avoid logging in server access logs)
 *     responses:
 *       200:
 *         description: Secret retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   type: object
 *                   properties:
 *                     secret:
 *                       type: string
 *       404:
 *         description: Secret not found
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/NotFoundResponse'
 *       403:
 *         description: Forbidden (expired, view limit reached, IP not allowed, or invalid secondary secret)
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/ForbiddenResponse'
 *       500:
 *         description: Server error
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/InternalServerErrorResponse'
 */
router.get(
  "/:identifier",
  strictLimiter,
  validateSecret,
  async (
    req: Request<{ identifier: string }, {}, any>,
    res: Response<IApiResponse<IGetSecretResponse>>
  ) => {
    const secret = req.body.secret;
    res.send({ success: true, data: { secret: secret.encryptedSecret } });
  }
);

/**
 * @swagger
 * /api/secrets/{identifier}:
 *   delete:
 *     summary: Delete a given secret
 *     description: Delete a secret by its creator identifier. This action is irreversible.
 *     tags: [Secrets]
 *     parameters:
 *       - in: path
 *         name: identifier
 *         schema:
 *           type: string
 *         required: true
 *         description: The creator identifier of the secret
 *     responses:
 *       200:
 *         description: Secret deleted successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   type: object
 *                   properties:
 *                     message:
 *                       type: string
 *       404:
 *         description: Secret not found
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/NotFoundResponse'
 *       500:
 *         description: Server error
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/InternalServerErrorResponse'
 */
router.delete(
  "/:creatorIdentifier",
  async (req: Request, res: Response<IApiResponse<IDeleteSecretResponse>>) => {
    const { creatorIdentifier } = req.params;

    try {
      const secret = await Secret.findOneAndDelete({ creatorIdentifier });
      if (!secret) {
        return res
          .status(404)
          .send({ success: false, error: "Secret not found" });
      }
      res.send({ success: true, data: { message: "Secret deleted" } });
    } catch (error: any) {
      res.status(500).send({ success: false, error });
    }
  }
);

/**
 * @swagger
 * /api/secrets/logs/{identifier}:
 *   get:
 *     summary: Get the access logs for a secret
 *     description: Retrieve the access logs for a secret by its creator identifier
 *     tags: [Secrets]
 *     parameters:
 *       - in: path
 *         name: identifier
 *         schema:
 *           type: string
 *         required: true
 *         description: The creator identifier of the secret
 *     responses:
 *       200:
 *         description: Access logs retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   type: object
 *                   properties:
 *                     logs:
 *                       type: array
 *                       items:
 *                         $ref: '#/components/schemas/ISecretAccessLog'
 *       404:
 *         description: Secret not found
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/NotFoundResponse'
 */
router.get(
  "/logs/:creatorIdentifier",
  async (req: Request, res: Response<IApiResponse<IGetSecretLogsResponse>>) => {
    const { creatorIdentifier } = req.params;
    const secret = await Secret.findOne({
      creatorIdentifier: creatorIdentifier,
    });

    if (!secret) {
      return res
        .status(404)
        .send({ success: false, error: "Secret not found" });
    }

    res.send({ success: true, data: { logs: secret.accessLogs || [] } });
  }
);


/**
 * @swagger
 * /api/secrets/stats/{creatorIdentifier}:
 *   get:
 *     summary: Get usage statistics for a secret
 *     description: Retrieve dashboard-ready usage statistics for a secret by its creator identifier
 *     tags: [Secrets]
 *     parameters:
 *       - in: path
 *         name: creatorIdentifier
 *         schema:
 *           type: string
 *         required: true
 *         description: The creator identifier of the secret
 *     responses:
 *       200:
 *         description: Secret statistics retrieved successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 data:
 *                   $ref: '#/components/schemas/GetSecretStatsResponse'
 *       404:
 *         description: Secret not found
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/responses/NotFoundResponse'
 */
router.get(
    '/stats/:creatorIdentifier',
    async (req: Request, res: Response<IApiResponse<IGetSecretStatsResponse>>) => {
        const { creatorIdentifier } = req.params;
        const secret = await Secret.findOne({ creatorIdentifier });

        if (!secret) {
            return res.status(404).send({ success: false, error: 'Secret not found' });
        }

        const logs = secret.accessLogs ?? [];
        const maxViews = Number.isFinite(secret.maxViews) ? (secret.maxViews as number) : null;
        const expired = !!secret.expirationDate && new Date(secret.expirationDate) < new Date();
        const exhausted = maxViews !== null && (secret.currentViews ?? 0) >= maxViews;

        res.send({ success: true, data: {
            reportedViews: secret.currentViews ?? 0,
            totalViews: logs.length,
            uniqueViews: new Set(logs.map(l => l.ipAddress)).size,
            grantedAttempts: logs.filter(l => l.accessGranted).length,
            refusedAttempts: logs.filter(l => !l.accessGranted).length,
            maxViews,
            expirationDate: secret.expirationDate ? new Date(secret.expirationDate).toISOString() : null,
            createdAt: (secret._id as Types.ObjectId).getTimestamp().toISOString(),
            hasPassword: !!secret.secretPassword,
            ipRestrictions: secret.ipRestrictions ?? [],
            emailNotification: secret.emailNotification ?? null,
            status: expired ? 'expired' : exhausted ? 'exhausted' : 'active',
        }});
    }
)

export default router;